default lip sync fix + extended experimental lip sync feature

+++ tgcalls/src/video/PythonSource.cpp
@@ -1,7 +1,7 @@
 #include "PythonSource.h"
 
 
-PythonSource::PythonSource(std::function<std::string()> getNextFrameBuffer, int fps, int width, int height):
+PythonSource::PythonSource(std::function<std::string()> getNextFrameBuffer, float fps, int width, int height):
   _fps(fps), _width(width), _height(height) {
   _getNextFrameBuffer = std::move(getNextFrameBuffer);
 }
